+++ src/app.ts
@@ -40,9 +40,9 @@ checkEnvVars();
 // const fastify = Fastify({ logger: true });
 const fastify = Fastify({
     https: {
-            key: fs.readFileSync("/etc/ssl/private/key.pem"),
-            cert: fs.readFileSync("/etc/ssl/private/cert.pem")
-        }
+        key: fs.readFileSync("/etc/ssl/private/key.pem"),
+        cert: fs.readFileSync("/etc/ssl/private/cert.pem")
+    }
 });
 
 // Wichtig: Multipart muss VOR allen Routen registriert werden!
