--- src/backend/controllers/TicTacToeController.ts
@@ -1,4 +1,0 @@
-
-export class TicTacToeController {
-
-}
