+++ src/backend/services/EmailService.ts
@@ -1,64 +1,63 @@
 import nodemailer from 'nodemailer';
 import crypto from 'crypto';
-import dotenv from 'dotenv';
 
 interface EmailOptions {
-	to: string;
-	subject: string;
-	text: string;
-	html: string;
+    to: string;
+    subject: string;
+    text: string;
+    html: string;
 }
 
 export class EmailService {
-	private transporter: nodemailer.Transporter;
+    private transporter: nodemailer.Transporter;
 
-	constructor() {
-		// Create transporter using environment variables
-		this.transporter = nodemailer.createTransport({
-			host: process.env.EMAIL_HOST,
-			port: parseInt(process.env.EMAIL_PORT || '587', 10),
-			secure: process.env.EMAIL_PORT === '465', // true for 465, false for other ports
-			auth: {
-				user: process.env.EMAIL_USER,
-				pass: process.env.EMAIL_PASSWORD,
-			},
-		});
-	}
+    constructor() {
+        // Create transporter using environment variables
+        this.transporter = nodemailer.createTransport({
+            host: process.env.EMAIL_HOST,
+            port: parseInt(process.env.EMAIL_PORT || '587', 10),
+            secure: process.env.EMAIL_PORT === '465', // true for 465, false for other ports
+            auth: {
+                user: process.env.EMAIL_USER,
+                pass: process.env.EMAIL_PASSWORD,
+            },
+        });
+    }
 
-	// Generate a random token for verification or password reset
-	generateToken(): string {
-		return crypto.randomBytes(32).toString('hex');
-	}
+    // Generate a random token for verification or password reset
+    generateToken(): string {
+        return crypto.randomBytes(32).toString('hex');
+    }
 
-	// Send email using the configured transporter
-	async sendEmail(options: EmailOptions): Promise<void> {
-		try {
-			await this.transporter.sendMail({
-				from: `"Transcendence App" <${process.env.EMAIL_USER}>`,
-				to: options.to,
-				subject: options.subject,
-				text: options.text,
-				html: options.html,
-			});
-			console.log(`Email sent to ${options.to}`);
-		} catch (error) {
-			console.error('Error sending email:', error);
-			throw new Error('Failed to send email');
-		}
-	}
+    // Send email using the configured transporter
+    async sendEmail(options: EmailOptions): Promise<void> {
+        try {
+            await this.transporter.sendMail({
+                from: `"Transcendence App" <${process.env.EMAIL_USER}>`,
+                to: options.to,
+                subject: options.subject,
+                text: options.text,
+                html: options.html,
+            });
+            console.log(`Email sent to ${options.to}`);
+        } catch (error) {
+            console.error('Error sending email:', error);
+            throw new Error('Failed to send email');
+        }
+    }
 
-	// Send verification email for new account
-	async sendVerificationEmail(email: string, token: string, username: string): Promise<void> {
-		const baseUrl = process.env.BASE_URL || 'http://localhost:3000';
+    // Send verification email for new account
+    async sendVerificationEmail(email: string, token: string, username: string): Promise<void> {
+        const baseUrl = process.env.BASE_URL || 'http://localhost:3000';
 
-		// Use API endpoint for verification because it will redirect to login
-		const verificationLink = `${baseUrl}/api/verify-email/${token}`;
+        // Use API endpoint for verification because it will redirect to login
+        const verificationLink = `${baseUrl}/api/verify-email/${token}`;
 
-		await this.sendEmail({
-			to: email,
-			subject: 'Verify Your Transcendence Account',
-			text: `Hello ${username},\n\nPlease verify your account by clicking the link: ${verificationLink}\n\nIf you did not request this, please ignore this email.\n`,
-			html: `
+        await this.sendEmail({
+            to: email,
+            subject: 'Verify Your Transcendence Account',
+            text: `Hello ${username},\n\nPlease verify your account by clicking the link: ${verificationLink}\n\nIf you did not request this, please ignore this email.\n`,
+            html: `
                 <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto; background-color: #0a1a2f; color: white; padding: 20px; border-radius: 8px;">
                     <h2 style="color: white; text-align: center; margin-bottom: 30px;">Welcome to Transcendence!</h2>
                     <p>Hello ${username},</p>
@@ -70,21 +69,21 @@ export class EmailService {
                     <p style="margin-top: 30px; border-top: 1px solid #2a3a4f; padding-top: 15px;">Thank you,<br>The Transcendence Team</p>
                 </div>
             `
-		});
-	}
+        });
+    }
 
-	// Send password reset email
-	async sendPasswordResetEmail(email: string, token: string, username: string): Promise<void> {
-		const baseUrl = process.env.BASE_URL || 'http://localhost:3000';
+    // Send password reset email
+    async sendPasswordResetEmail(email: string, token: string, username: string): Promise<void> {
+        const baseUrl = process.env.BASE_URL || 'http://localhost:3000';
 
-		// Changed to use frontend route instead of API endpoint
-		const resetLink = `${baseUrl}/password-reset/${token}`;
+        // Changed to use frontend route instead of API endpoint
+        const resetLink = `${baseUrl}/password-reset/${token}`;
 
-		await this.sendEmail({
-			to: email,
-			subject: 'Password Reset for Transcendence',
-			text: `Hello ${username},\n\nYou requested a password reset. Please click the following link to reset your password: ${resetLink}\n\nThis link will expire in 1 hour.\n\nIf you did not request this, please ignore this email.\n`,
-			html: `
+        await this.sendEmail({
+            to: email,
+            subject: 'Password Reset for Transcendence',
+            text: `Hello ${username},\n\nYou requested a password reset. Please click the following link to reset your password: ${resetLink}\n\nThis link will expire in 1 hour.\n\nIf you did not request this, please ignore this email.\n`,
+            html: `
                 <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto; background-color: #0a1a2f; color: white; padding: 20px; border-radius: 8px;">
                     <h2 style="color: white; text-align: center; margin-bottom: 30px;">Password Reset Request</h2>
                     <p>Hello ${username},</p>
@@ -97,6 +96,6 @@ export class EmailService {
                     <p style="margin-top: 30px; border-top: 1px solid #2a3a4f; padding-top: 15px;">Thank you,<br>The Transcendence Team</p>
                 </div>
             `
-		});
-	}
+        });
+    }
 }
