+++ src/frontend/components/Footer.ts
@@ -1,12 +1,12 @@
 import AbstractView from '../../utils/AbstractView.js';
 
 export default class Footer extends AbstractView {
-	constructor(params: URLSearchParams = new URLSearchParams(window.location.search)) {
-		super(params);
-	}
+    constructor(params: URLSearchParams = new URLSearchParams(window.location.search)) {
+        super(params);
+    }
 
-	async getHtml(): Promise<string> {
-		return super.render(`
+    async getHtml(): Promise<string> {
+        return super.render(`
 			<footer class="w-full py-1 px-5">
 				<div class="flex flex-col md:flex-row justify-between items-center text-sm w-full">
 					<p>&copy; {{props.year || '2025'}} Transcendence Project</p>
@@ -22,5 +22,5 @@ export default class Footer extends AbstractView {
 				</div>
 			</footer>
 		`);
-	}
+    }
 }
