+++ src/frontend/components/Label.ts
@@ -1,35 +1,26 @@
-// ========================
-// File: components/Label.ts
-// ========================
-
 import AbstractView from '../../utils/AbstractView.js';
-
-interface LabelProps {
-	htmlFor: string;
-	text: string;
-	className?: string;
-}
+import { LabelProps } from '../../interfaces/abstractViewInterfaces.js';
 
 export default class Label extends AbstractView {
-	constructor(params: URLSearchParams = new URLSearchParams()) {
-		super(params);
-	}
+    constructor(params: URLSearchParams = new URLSearchParams()) {
+        super(params);
+    }
 
-	async renderLabel({ htmlFor, text, className = '' }: LabelProps): Promise<string> {
-		const labelClass = className
+    async renderLabel({ htmlFor, text, className = '' }: LabelProps): Promise<string> {
+        const labelClass = className
 
-		return this.render(`
+        return this.render(`
 			<label for="${htmlFor}" class="${labelClass}">
 				${text}
 			</label>
 		`);
-	}
+    }
 
-	async getHtml(): Promise<string> {
-		return this.render(`
+    async getHtml(): Promise<string> {
+        return this.render(`
 			<label for="default">
 				Default Label
 			</label>
 		`);
-	}
+    }
 }
