+++ src/frontend/views/ProfileEdit.ts
@@ -2,54 +2,54 @@ import Title from '../components/Title.js';
 import Card from '../components/Card.js';
 import Button from '../components/Button.js';
 import AbstractView from '../../utils/AbstractView.js';
-import { generateProfileImage } from '../../utils/Avartar.js';
+import { generateProfileImage } from '../../utils/Avatar.js';
 
 export default class ProfileEdit extends AbstractView {
-	private userId: string;
+    private userId: string;
 
-	constructor(params: URLSearchParams) {
-		super();
-		this.userId = params.get('id') || 'unknown';
-	}
+    constructor(params: URLSearchParams) {
+        super();
+        this.userId = params.get('id') || 'unknown';
+    }
 
-	async getHtml(): Promise<string> {
+    async getHtml(): Promise<string> {
 
-		// User data will be stored here
-		let userData = null;
-		let users = [];
+        // User data will be stored here
+        let userData = null;
+        let users = [];
 
-		try {
-			// First, fetch the specific user using the userId from the route parameters
-			const userResponse = await fetch(`/api/users/${this.userId}`);
-			if (userResponse.ok) {
-				userData = await userResponse.json();
-				console.log('User data loaded:', userData);
-			} else {
-				console.error('Failed to fetch user data from API');
-			}
-		} catch (error) {
-			console.error('API request error:', error);
-		}
+        try {
+            // First, fetch the specific user using the userId from the route parameters
+            const userResponse = await fetch(`/api/users/${this.userId}`);
+            if (userResponse.ok) {
+                userData = await userResponse.json();
+                console.log('User data loaded:', userData);
+            } else {
+                console.error('Failed to fetch user data from API');
+            }
+        } catch (error) {
+            console.error('API request error:', error);
+        }
 
-		// Title section
-		const title = new Title({
-			title: userData ? `Profile: ${userData.displayname}` : 'User Profile',
-		});
-		const titleSection = await title.getHtml();
+        // Title section
+        const title = new Title({
+            title: userData ? `Profile: ${userData.displayname}` : 'User Profile',
+        });
+        const titleSection = await title.getHtml();
 
-		// User profile card
-		const card = new Card();
+        // User profile card
+        const card = new Card();
 
-		let profileCardHtml = '';
-		let profileImageSvg = '';
+        let profileCardHtml = '';
+        let profileImageSvg = '';
 
-		if (userData) {
-			// Generate the profile image SVG
-			profileImageSvg = generateProfileImage(userData, 200, 200);
+        if (userData) {
+            // Generate the profile image SVG
+            profileImageSvg = generateProfileImage(userData, 200, 200);
 
-			profileCardHtml = await card.renderCard({
-				title: 'User Profile',
-				extra: `
+            profileCardHtml = await card.renderCard({
+                title: 'User Profile',
+                extra: `
                 <div class="profile-container">
                     <div class="profile-header">
                         <div class="profile-avatar-container">
@@ -70,37 +70,37 @@ export default class ProfileEdit extends AbstractView {
                         <!-- Add more user details as needed -->
                     </div>
                 </div>`,
-				data: { user: userData }
-			});
-		} else {
-			profileCardHtml = await card.renderCard({
-				title: 'User Profile',
-				extra: `<div class="alert alert-warning">User not found or error loading user data.</div>`
-			});
-		}
+                data: { user: userData }
+            });
+        } else {
+            profileCardHtml = await card.renderCard({
+                title: 'User Profile',
+                extra: `<div class="alert alert-warning">User not found or error loading user data.</div>`
+            });
+        }
 
-		// Button for edit
-		const button = new Button();
-		const buttonGroup = await button.renderGroup({
-			layout: 'stack',
-			align: 'center',
-			buttons: [
-				{
-					id: 'back-to-list',
-					text: 'Back to User List',
-					href: '/user-mangement',
-					className: 'btn btn-primary'
-				}
-			]
-		});
+        // Button for edit
+        const button = new Button();
+        const buttonGroup = await button.renderGroup({
+            layout: 'stack',
+            align: 'center',
+            buttons: [
+                {
+                    id: 'back-to-list',
+                    text: 'Back to User List',
+                    href: '/user-mangement',
+                    className: 'btn btn-primary'
+                }
+            ]
+        });
 
-		// Final output
-		return this.render(`
+        // Final output
+        return this.render(`
             <div class="container">
                 ${titleSection}
                 ${profileCardHtml}
                 ${buttonGroup}
             </div>
         `);
-	}
+    }
 }
