+++ src/frontend/views/UserManagement.ts
@@ -1,98 +1,88 @@
 import Title from '../components/Title.js';
 import Card from '../components/Card.js';
 import Button from '../components/Button.js';
-import { generateProfileImage } from '../../utils/Avartar.js';
+import { generateProfileImage } from '../../utils/Avatar.js';
 import AbstractView from '../../utils/AbstractView.js';
-
-interface UserList {
-	listAvatar: string;
-	avatar: string;
-	id?: number;
-	username: string;
-	email: string;
-	displayname?: string;
-	password?: string;
-	role?: string;
-}
+import { UserList } from '../../interfaces/userInterfaces.js';
 
 export default class UserManagement extends AbstractView {
-	constructor() {
-		super();
-	}
+    constructor() {
+        super();
+    }
 
-	async getHtml(): Promise<string> {
+    async getHtml(): Promise<string> {
 
-		// Fetch users from API
-		let users = [];
-		try {
-			const response = await fetch('/api/users/');
-			if (response.ok) {
-				users = await response.json();
-			} else {
-				console.error('Failed to fetch users from API');
-			}
-		} catch (error) {
-			console.error('API request error:', error);
-		}
+        // Fetch users from API
+        let users = [];
+        try {
+            const response = await fetch('/api/users/');
+            if (response.ok) {
+                users = await response.json();
+            } else {
+                console.error('Failed to fetch users from API');
+            }
+        } catch (error) {
+            console.error('API request error:', error);
+        }
 
-		// Correctly loop through each user and add avatar
-		users.forEach((user: UserList) => {
-			user.listAvatar = generateProfileImage(user, 20, 20);
-		});
+        // Correctly loop through each user and add avatar
+        users.forEach((user: UserList) => {
+            user.listAvatar = generateProfileImage(user, 20, 20);
+        });
 
-		// Title section
-		const title = new Title({
-			title: 'User Management',
-		});
-		const titleSection = await title.getHtml();
+        // Title section
+        const title = new Title({
+            title: 'User Management',
+        });
+        const titleSection = await title.getHtml();
 
-		// Button Group (uses  automatically)
-		const button = new Button();
-		const readAllButtonGroup = await button.renderGroup({
-			layout: 'stack',
-			align: 'center',
-			buttons: [
-				{
-					id: 'read-all',
-					text: 'Read all Users',
-					onClick: `document.getElementById('user-list').innerHTML = '<li>Loading...</li>'`
-				}
-			]
-		});
+        // Button Group (uses  automatically)
+        const button = new Button();
+        const readAllButtonGroup = await button.renderGroup({
+            layout: 'stack',
+            align: 'center',
+            buttons: [
+                {
+                    id: 'read-all',
+                    text: 'Read all Users',
+                    onClick: `document.getElementById('user-list').innerHTML = '<li>Loading...</li>'`
+                }
+            ]
+        });
 
-		// CRUD Form Cards
-		const card = new Card();
+        // CRUD Form Cards
+        const card = new Card();
 
-		const cardConfigs = [
-			// {
-			// 	title: 'Read One User',
-			// 	formId: 'read-one-form',
-			// 	inputs: [{ name: 'userId', type: 'number', placeholder: 'User ID' }],
-			// 	button: { text: 'Read User', type: 'submit' },
-			// 	extra: `<div id="read-result" class="text-white text-sm pt-2"></div>`
-			// },
-			// {
-			// 	title: 'Update User',
-			// 	formId: 'update-form',
-			// 	inputs: [
-			// 		{ name: 'id', type: 'number', placeholder: 'User ID' },
-			// 		{ name: 'name', placeholder: 'New Name' },
-			// 		{ name: 'username', placeholder: 'New Username' }
-			// 	],
-			// 	button: { text: 'Update', type: 'submit' }
-			// },
-			// {
-			// 	title: 'Delete User',
-			// 	formId: 'delete-form',
-			// 	inputs: [{ name: 'id', type: 'number', placeholder: 'User ID' }],
-			// 	button: { text: 'Delete', type: 'submit', className: 'btn btn-danger btn-sm' },
-			// }
-		];
+        const cardConfigs = [
+            // {
+            // 	title: 'Read One User',
+            // 	formId: 'read-one-form',
+            // 	inputs: [{ name: 'userId', type: 'number', placeholder: 'User ID' }],
+            // 	button: { text: 'Read User', type: 'submit' },
+            // 	extra: `<div id="read-result" class="text-white text-sm pt-2"></div>`
+            // },
+            // {
+            // 	title: 'Update User',
+            // 	formId: 'update-form',
+            // 	inputs: [
+            // 		{ name: 'id', type: 'number', placeholder: 'User ID' },
+            // 		{ name: 'name', placeholder: 'New Name' },
+            // 		{ name: 'username', placeholder: 'New Username' }
+            // 	],
+            // 	button: { text: 'Update', type: 'submit' }
+            // },
+            // {
+            // 	title: 'Delete User',
+            // 	formId: 'delete-form',
+            // 	inputs: [{ name: 'id', type: 'number', placeholder: 'User ID' }],
+            // 	button: { text: 'Delete', type: 'submit', className: 'btn btn-danger btn-sm' },
+            // }
+        ];
 
-		// List Card
-		const listCard = await card.renderCard({
-			title: 'Users',
-			extra: `<table class="list" data-height="400px">
+        // List Card
+        const listCard = await card.renderCard({
+            title: 'Users',
+            extra: `<table class="list" data-height="400px">
 				<thead>
 				<tr>
 					<th>Avatar</th>
@@ -126,44 +116,44 @@ export default class UserManagement extends AbstractView {
 				</for>
 				</tbody>
 			</table>`,
-			data: { users }
-		});
+            data: { users }
+        });
 
-		// Register Card
-		const registerCard = await card.renderCard({
-			title: 'Create User',
-			formId: 'create-form',
-			inputs: [
-				{ name: 'displayname', type: "text", placeholder: 'Name' },
-				{ name: 'username', type: "text", placeholder: 'Username' },
-				{ name: 'email', type: 'email', placeholder: 'Email Address' },
-				{
-					name: 'role',
-					type: 'select',
-					placeholder: 'Select Role',
-					options: [
-						{
-							label: "User",
-							value: "user"
-						},
-						{
-							label: "Admin",
-							value: "admin"
-						}
-					]
-				},
-				{ name: 'password', type: 'password', placeholder: 'Password' }
-			],
-			button: { text: 'Create', type: 'submit', className: "btn btn-primary" }
-		});
+        // Register Card
+        const registerCard = await card.renderCard({
+            title: 'Create User',
+            formId: 'create-form',
+            inputs: [
+                { name: 'displayname', type: "text", placeholder: 'Name' },
+                { name: 'username', type: "text", placeholder: 'Username' },
+                { name: 'email', type: 'email', placeholder: 'Email Address' },
+                {
+                    name: 'role',
+                    type: 'select',
+                    placeholder: 'Select Role',
+                    options: [
+                        {
+                            label: "User",
+                            value: "user"
+                        },
+                        {
+                            label: "Admin",
+                            value: "admin"
+                        }
+                    ]
+                },
+                { name: 'password', type: 'password', placeholder: 'Password' }
+            ],
+            button: { text: 'Create', type: 'submit', className: "btn btn-primary" }
+        });
 
-		// Final output - Pass users data to the render method
-		return this.render(`
+        // Final output - Pass users data to the render method
+        return this.render(`
 			<div class="container">
 				${titleSection}
 				${registerCard}
 				${listCard}
 			</div>
 		`);
-	}
+    }
 }
