+++ src/interfaces/interfaces.ts
@@ -25,24 +25,6 @@ export interface IGameState {
     gameIsOver: boolean;
 }
 
-export interface User {
-	id?: number;
-	username: string;
-	email: string;
-	displayname?: string;
-	password?: string;
-	role?: string;
-	avatar?: string;
-	emailVerified?: boolean;
-	tf_one?: string;
-	tf_two?: string;
-	tf_three?: string;
-	tf_four?: string;
-	tf_five?: string;
-	tf_six?: string;
-	secret?: string;
-}
-
 export interface LobbyInfo{
     id: string;
     name: string;
