+++ src/utils/checkEnvVars.ts
@@ -1,15 +1,15 @@
 export default function checkEnvVars() {
-	const requiredVars: string[] = ['JWT_SECRET', 'COOKIE_SECRET', 'REFRESH_TOKEN_SECRET'];
-	const missingVars: string[] = requiredVars.filter(varName => {
-		!process.env[varName];
-	});
+    const requiredVars: string[] = ['JWT_SECRET', 'COOKIE_SECRET', 'REFRESH_TOKEN_SECRET'];
+    const missingVars: string[] = requiredVars.filter(varName => {
+        !process.env[varName];
+    });
 
-	if (missingVars.length > 0) {
-		console.error(`Missing required environment variables: ${missingVars.join(', ')}`);
-		console.error("Please check your .env file and make sure it is being loaded correctly.");
-		process.exit(1);
-	}
-	else {
-		console.log("All required environment variables are loaded.");
-	}
+    if (missingVars.length > 0) {
+        console.error(`Missing required environment variables: ${missingVars.join(', ')}`);
+        console.error("Please check your .env file and make sure it is being loaded correctly.");
+        process.exit(1);
+    }
+    else {
+        console.log("All required environment variables are loaded.");
+    }
 }
